submitorder in shopping cart

+++ HHMarketWebApp/HHMarketWebApp/Models/ProductionDetail.cs
@@ -20,7 +20,7 @@ namespace HHMarketWebApp.Models
 
         public decimal Price { get; set; }
 
-        public short Amount { get; set; }
+        public int Amount { get; set; }
 
         public int ProductId { get; set; }
